Create Linkage File of SM&AC
This is the result file linked in FRIL. In linking the data of SM and
AC, I set the params as follows:

[1]Artists' Name: approve level is 0.1, disapprove level is 0.45, and
its weight is 50.

[2]Artists' birthYear: approve level is 0.05, disapprove level is 0.3,
and its weight is 25.

[3]Artists' deathYear: approve level is 0.05, disapprove level is 0.3,
and its weight is 25.

[4]The acceptance level: 51.

I tested other params and the above was perfect. With the above
configuration the FRIL links 91 records.
There are more result than last time, but a little linkage maybe not
accurate when I check each record.